requirements.txt
Django==1.7.2
Markdown==2.5.2
django-filter==0.9.1
djangorestframework==3.0.2
psycopg2==2.5.4
requests==2.5.0
selenium==2.44.0
